Don't block status collection
Our current status reporter works as a part of the event loop. This
design has a downside that the status collection is blocked when the
event loop is blocked or stopped. Since we want to detect that state as
well, we should decouple the status collection from the event loop.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>