Wait for IPv6 Link Local Address to be assigned
In the previous commit, we replace the retrying logic of socket creation
to the netlink-based device detection. The idea was waiting on the
device to be appeared and up so that the socket creation will always
succeed.

However, that was not enough. In Linux, there's a small time window
which the device is up but the IPv6 link-local address is not assigned.
If we create the socket within this window, it fails because it calls
bind(2) internally and it depends on the link-local address to be
present.

To solve this issue, wait on the address event of the netlink and ensure
the link-local address is assigned before creating socket.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>